"Bogmaw's Claw" Quest Reward Fix
Reported by Kierkan (kierkan) in Discord.
(https://discord.com/channels/1043651040962158642/1043651040962158645/1370969983529779355)

Maybe because it's referencing icon 2 times, the spell fails to initialize and can't be used.

I'm assuming that "icon: @_icon" is the syntax to reference the "_icon" listed above in "item .swamp_giant_1_a", which means an icon shouldn't need to be set again within the ability.